Updated to process all edge types.
Now EdgeListReverseIndex will submit a MapReduce job for each edge type in the
graph. Each job will consume all the edge list files for that edge type and
produce a reverse index in a separate output folder, `output/edgeType`.

Also hardcoded the number of reducers to 9, which needs to be set dynamically in
the future, but it's not clear how to query the number of nodes in the cluster
before submitting the job.